Fixed traits and Auth

+++ app/Http/Controllers/UsersController.php
@@ -124,10 +124,13 @@ class UsersController extends Controller
     public function destroy(User $user)
     {
         if(Auth::user()->user_type != 2){
-            return $this->restricted('', 'Nije vam dozvoljeno!', 403);
+            return $this->restricted('', 'Nije vam dozvoljeno brisanje!', 403);
         }
         if($user->user_type == 2) {
             return $this->restricted('', 'Nije vam dozvoljeno da obrišete Administratora!', 403);
+        }
+        if($user->user_type == 0 && $user->club != null) {
+            return $this->restricted('', 'Morate premjestiti klub ili obrisati kako bi ste obrisali korisnika!', 403);
         }
         $user->delete();
